Extra modes for SimValidator
By default, this validator now accepts all kinds of SIM card numbers, with or without check digits.
The check digit is only validated when it is sure that there is one (20 digits numbers).

The validator is also configurable to accommodate use cases when it is known for sure whether the number contains a check digit or not.